split gulp into modules
